ソング：シーケンサーのグリッドをコンポーネント化する (#2087)
* グリッドをコンポーネント化した

* 最後の小節でマウスホイールで縮小を行ったときに表示がおかしくなるのを修正

* PreviewModeの文字列を変更
